+++ lua/xmap/lang/cc.lua
@@ -1,0 +1,3 @@
+-- lua/xmap/lang/cc.lua
+-- `cc` is treated as a C++ synonym and reuses the C++ provider.
+return require("xmap.lang.cpp")